id -> username

+++ src/main/java/ru/binarysimple/users/controller/UserController.java
@@ -4,7 +4,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.binarysimple.users.dto.CreateUserDto;
 import ru.binarysimple.users.dto.UserDto;
 import ru.binarysimple.users.service.UserService;
@@ -18,7 +20,6 @@ import java.io.IOException;
 public class UserController {
 
     private final UserService userService;
-
 
 //    @GetMapping(params = {"id"})
 //    public UserDto getOne(@RequestParam Long id) {
@@ -42,17 +43,26 @@ public class UserController {
 //    }
 
     @GetMapping(params = {"username"})
-    public UserDto getOneByUsername(@RequestParam String username) {
+    public UserDto getOneByUsername(@RequestHeader("X-Username") String currentUsername, @RequestParam String username) {
+        if (!currentUsername.equals(username)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        }
         return userService.getByUsername(username);
     }
 
     @PutMapping(params = {"username"})
-    public UserDto updateByUsername(@RequestParam String username, @RequestBody JsonNode patchNode) throws IOException {
+    public UserDto updateByUsername(@RequestHeader("X-Username") String currentUsername, @RequestParam String username, @RequestBody JsonNode patchNode) throws IOException {
+        if (!currentUsername.equals(username)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        }
         return userService.updateByUsername(username, patchNode);
     }
 
     @DeleteMapping(params = {"username"})
-    public UserDto delete(@RequestParam String username) {
+    public UserDto delete(@RequestHeader("X-Username") String currentUsername, @RequestParam String username) {
+        if (!currentUsername.equals(username)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        }
         return userService.deleteByUsername(username);
     }
 }
